requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
streamlit==1.28.1
transformers==4.35.2
torch==2.6.0
torchvision==0.21.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.1
numpy==1.26.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
jinja2==3.1.2
